+++ client/src/components/panel.js
@@ -1,36 +1,11 @@
 import React, {Component, PureComponent} from 'react';
 import {Link} from 'react-router-dom';
 import Translate from 'react-translate-component';
+import downloadImgFromMapCanvas from '../lib/map-to-img.js';
 
 class PanelHeader extends PureComponent {
   downloadImg() {
-    const format = "image/png";
-    const quality = 0.9;
-
-    // Create copy
-    const mapCanvas = this.props.map.getCanvas();
-    const copy = document.createElement('canvas');
-    copy.width = mapCanvas.width;
-    copy.height = mapCanvas.height;
-    const copyCtx = copy.getContext('2d');
-    copyCtx.drawImage(mapCanvas, 0, 0);
-
-    // Add watermark to copy
-    const watermark = `citylines.co/${this.props.urlName}`;
-    copyCtx.font = "35px Arial";
-    const width = copyCtx.measureText(watermark).width;
-    copyCtx.fillText(watermark, copy.width - width - 90, copy.height - 70);
-
-    copy.toBlob(blob => {
-      const anchor = document.createElement('a');
-      anchor.download = `${this.props.urlName}.png`;
-      anchor.href = URL.createObjectURL(blob);
-      anchor.click();
-      URL.revokeObjectURL(anchor.href);
-    },
-    format,
-    quality,
-    );
+    downloadImgFromMapCanvas(this.props.urlName, this.props.map.getCanvas());
   }
 
   render() {
